\name{ImageSegment-package}
\alias{ImageSegment-package}
\alias{ImageSegment}
\docType{package}
\title{
\packageTitle{ImageSegment}
}
\description{
\packageDescription{ImageSegment}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{ImageSegment}
\packageIndices{ImageSegment}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{ImageSegment}

Maintainer: \packageMaintainer{ImageSegment}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R ~~
~~ documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
